+++ repos/ktor/client/src/commonMain/kotlin/dev/inmo/micro_utils/repos/ktor/client/key/values/KtorKeyValuesRepoClient.kt
@@ -1,10 +1,15 @@
 package dev.inmo.micro_utils.repos.ktor.client.key.values
 
+import dev.inmo.micro_utils.ktor.client.createStandardWebsocketFlow
 import dev.inmo.micro_utils.ktor.common.*
 import dev.inmo.micro_utils.repos.*
+import dev.inmo.micro_utils.repos.ktor.common.one_to_many.onDataClearedRoute
+import dev.inmo.micro_utils.repos.ktor.common.one_to_many.onNewValueRoute
+import dev.inmo.micro_utils.repos.ktor.common.one_to_many.onValueRemovedRoute
 import io.ktor.client.HttpClient
 import io.ktor.http.ContentType
 import io.ktor.http.encodeURLQueryComponent
+import kotlinx.coroutines.flow.Flow
 import kotlinx.serialization.*
 
 class KtorKeyValuesRepoClient<Key, Value> (
@@ -20,6 +25,15 @@ class KtorKeyValuesRepoClient<Key, Value> (
             httpClient: HttpClient,
             contentType: ContentType,
             noinline keySerializer: suspend (Key) -> String,
+            onNewValue: Flow<Pair<Key, Value>> = httpClient.createStandardWebsocketFlow(
+                buildStandardUrl(baseUrl, onNewValueRoute),
+            ),
+            onValueRemoved: Flow<Pair<Key, Value>> = httpClient.createStandardWebsocketFlow(
+                buildStandardUrl(baseUrl, onValueRemovedRoute),
+            ),
+            onDataCleared: Flow<Key> = httpClient.createStandardWebsocketFlow(
+                buildStandardUrl(baseUrl, onDataClearedRoute),
+            ),
             noinline valueSerializer: suspend (Value) -> String
         ) = KtorKeyValuesRepoClient(
             KtorReadKeyValuesRepoClient(
@@ -32,7 +46,10 @@ class KtorKeyValuesRepoClient<Key, Value> (
             KtorWriteKeyValuesRepoClient(
                 baseUrl,
                 httpClient,
-                contentType
+                contentType,
+                onNewValue,
+                onValueRemoved,
+                onDataCleared
             )
         )
         inline operator fun <reified Key : Any, reified Value : Any> invoke(
@@ -41,12 +58,24 @@ class KtorKeyValuesRepoClient<Key, Value> (
             httpClient: HttpClient,
             contentType: ContentType,
             noinline keySerializer: suspend (Key) -> String,
+            onNewValue: Flow<Pair<Key, Value>> = httpClient.createStandardWebsocketFlow(
+                buildStandardUrl(baseUrl, onNewValueRoute),
+            ),
+            onValueRemoved: Flow<Pair<Key, Value>> = httpClient.createStandardWebsocketFlow(
+                buildStandardUrl(baseUrl, onValueRemovedRoute),
+            ),
+            onDataCleared: Flow<Key> = httpClient.createStandardWebsocketFlow(
+                buildStandardUrl(baseUrl, onDataClearedRoute),
+            ),
             noinline valueSerializer: suspend (Value) -> String
         ) = KtorKeyValuesRepoClient(
             buildStandardUrl(baseUrl, subpart),
             httpClient,
             contentType,
             keySerializer,
+            onNewValue,
+            onValueRemoved,
+            onDataCleared,
             valueSerializer
         )
     }
@@ -59,13 +88,25 @@ inline fun <reified Key : Any, reified Value : Any> KtorKeyValuesRepoClient(
     keySerializer: SerializationStrategy<Key>,
     valueSerializer: SerializationStrategy<Value>,
     serialFormat: StringFormat,
+    onNewValue: Flow<Pair<Key, Value>> = httpClient.createStandardWebsocketFlow(
+        buildStandardUrl(baseUrl, onNewValueRoute),
+    ),
+    onValueRemoved: Flow<Pair<Key, Value>> = httpClient.createStandardWebsocketFlow(
+        buildStandardUrl(baseUrl, onValueRemovedRoute),
+    ),
+    onDataCleared: Flow<Key> = httpClient.createStandardWebsocketFlow(
+        buildStandardUrl(baseUrl, onDataClearedRoute),
+    ),
 ) = KtorKeyValuesRepoClient<Key, Value>(
     baseUrl,
     httpClient,
     contentType,
     {
         serialFormat.encodeToString(keySerializer, it).encodeURLQueryComponent()
-    }
+    },
+    onNewValue,
+    onValueRemoved,
+    onDataCleared
 ) {
     serialFormat.encodeToString(valueSerializer, it).encodeURLQueryComponent()
 }
@@ -77,13 +118,25 @@ inline fun <reified Key : Any, reified Value : Any> KtorKeyValuesRepoClient(
     keySerializer: SerializationStrategy<Key>,
     valueSerializer: SerializationStrategy<Value>,
     serialFormat: BinaryFormat,
+    onNewValue: Flow<Pair<Key, Value>> = httpClient.createStandardWebsocketFlow(
+        buildStandardUrl(baseUrl, onNewValueRoute),
+    ),
+    onValueRemoved: Flow<Pair<Key, Value>> = httpClient.createStandardWebsocketFlow(
+        buildStandardUrl(baseUrl, onValueRemovedRoute),
+    ),
+    onDataCleared: Flow<Key> = httpClient.createStandardWebsocketFlow(
+        buildStandardUrl(baseUrl, onDataClearedRoute),
+    ),
 ) = KtorKeyValuesRepoClient<Key, Value>(
     baseUrl,
     httpClient,
     contentType,
     {
         serialFormat.encodeHex(keySerializer, it)
-    }
+    },
+    onNewValue,
+    onValueRemoved,
+    onDataCleared
 ) {
     serialFormat.encodeHex(valueSerializer, it)
 }